# docs/tasking-mvp/tasking-mvp.openapi.yaml
openapi: 3.0.3
info:
  title: Workspaces Tasking Manager API
  version: 1.0.0
  description: |
    Tasking Manager API for the Workspaces Tasking Manager MVP.
servers:
  - url: /api/v1
    description: workspaces-backend FastAPI app — all routers are mounted under `/api/v1` (see `api/main.py`).
...
security:
  - bearerAuth: []
paths:
...
  /workspaces/{workspace_id}/tasking/projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
    get:
      tags: [projects]
      summary: Get a tasking project.
      operationId: getProject
      responses:
        "200":
          description: Project detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ProjectOrWorkspaceNotFound" }
    patch:
      tags: [projects]
      summary: Update editable settings.
      description: |
        LEAD only. Per-status mutability:

        | Field              | draft | open | done |
        |--------------------|-------|------|------|
        | name               | yes   | yes  | no   |
        | instructions       | yes   | yes  | no   |
        | lockTimeoutHours   | yes   | yes  | no   |
        | reviewRequired     | yes   | NO   | no   |

        AOI is updated through the dedicated AOI endpoints, not PATCH.
        Immutable-field writes return 422.
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdateRequest" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ForbiddenLeadRequired" }
        "404": { $ref: "#/components/responses/ProjectOrWorkspaceNotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      tags: [projects]
      summary: Soft-delete a tasking project.
      description: |
        LEAD only. Sets `deletedAt`, hard-deletes child tasks, retains
        audit rows (flagged with `project_deleted=true`). Returns 409
        if any task currently has an active lock — force-release first.
      operationId: deleteProject
      responses:
        "204":
          description: Soft-deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ForbiddenLeadRequired" }
        "404": { $ref: "#/components/responses/ProjectOrWorkspaceNotFound" }
        "409":
          description: Project has active task locks.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /workspaces/{workspace_id}/tasking/projects/{project_id}/tasks/validate:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
    post:
      tags: [tasks]
      summary: Validate a GeoJSON FeatureCollection of candidate task boundaries.
      description: |
        LEAD only. Does NOT persist anything — the client must POST each
        validated Feature to `/tasks/save` to commit.

        Project preconditions: status `draft`, AOI uploaded.

        Hard validation (422 on failure):
          - Top-level type must be `FeatureCollection`.
          - Every feature geometry must be `Polygon` (no MultiPolygon).
          - Coordinates valid lon/lat (EPSG:4326).
          - Every polygon lies within the project AOI (centroid-inside
            is not sufficient).
          - At least one feature.

        Non-blocking warning:
          - `polygon_exceeds_grid_size` — area exceeds
            `TM_TASKING_GRID_SIZE_METERS`² (km²).
      operationId: validateTasks
      requestBody:
        required: true
        content:
          application/geo+json:
            schema:
              { $ref: "#/components/schemas/TaskBoundariesFeatureCollection" }
          application/json:
            schema:
              { $ref: "#/components/schemas/TaskBoundariesFeatureCollection" }
      responses:
        "200":
          description: Validation result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidatePreviewResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ForbiddenLeadRequired" }
        "404": { $ref: "#/components/responses/ProjectOrWorkspaceNotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /workspaces/{workspace_id}/tasking/projects/{project_id}/tasks/save:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
      - $ref: "#/components/parameters/IdempotencyKeyHeader"
    post:
      tags: [tasks]
      summary: Persist a validated FeatureCollection as the project's tasks (bulk).
      description: |
        LEAD only. Commits the **entire** FeatureCollection as
        `tasking_tasks` rows in a single transaction.

        Effect (single transaction; all-or-nothing):
          1. Re-validates every feature against the project AOI (same
             rules as `POST /tasks/validate`). Any hard failure aborts
             the transaction — no rows are written.
          2. Allocates sequential `taskNumber`s starting at 1, in the
             order features appear in the request.
          3. Bulk-inserts the `tasking_tasks` rows.
          4. Sets `tasking_projects.task_boundary_type` to `source`.
          5. Emits one `task_created` audit event per task.

        Project preconditions:
          - Project must be in `draft`.
          - Project must have an AOI set.
          - Project must currently have **no tasks** (a second save is
            rejected with 409 `"Tasks already saved"`). To replace
            tasks, re-upload the AOI in `draft` — that wipes existing
            tasks and re-enables save.

        Idempotency: `Idempotency-Key` is honoured at the batch level.
        Same key + same body within the configured TTL replays the
        original response (HTTP 200, `replayed: true`). Same key with
        different body returns 409
        `"Idempotency key reused with a different request"`. See the
        top-level idempotency contract.
      operationId: saveTasks
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaveTasksRequest" }
      responses:
        "201":
          description: Tasks created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveTasksResponse" }
        "200":
          description: |
            Idempotent replay (same key + same body); `replayed: true`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveTasksResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ForbiddenLeadRequired" }
        "404": { $ref: "#/components/responses/ProjectOrWorkspaceNotFound" }
        "409":
          description: |
            One of:
              - "Tasks already saved" — project already has tasks; re-upload AOI to start over.
              - "Idempotency key reused with a different request".
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /workspaces/{workspace_id}/tasking/projects/{project_id}/tasks:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
    get:
      tags: [tasks]
      summary: List tasks (geometry always included; serves list + map views).
      description: |
        Paginated. Any workspace contributor.
      operationId: listTasks
      parameters:
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/TaskStatus" }
        - in: query
          name: lockedByUserId
          schema: { type: string, format: uuid }
        - in: query
          name: lastMapperId
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        "200":
          description: Tasks.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ProjectOrWorkspaceNotFound" }
  /workspaces/{workspace_id}/tasking/projects/{project_id}/tasks/{task_number}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
      - $ref: "#/components/parameters/TaskNumberPath"
    get:
      tags: [tasks]
      summary: Get task detail (geometry, status, lock, last mapper).
      operationId: getTask
      responses:
        "200":
          description: Task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Workspace, project, or task not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # =========================================================================
  # Locking + Submit
  # =========================================================================
  /workspaces/{workspace_id}/tasking/projects/{project_id}/tasks/{task_number}/lock:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
      - $ref: "#/components/parameters/TaskNumberPath"
    post:
      tags: [locks]
      summary: Acquire a lock on a task.
      description: |
        Eligibility:

        | Task status | Allowed roles                                     |
        |-------------|---------------------------------------------------|
        | `to_map`    | CONTRIBUTOR, LEAD                                 |
        | `to_remap`  | CONTRIBUTOR, LEAD                                 |
        | `to_review` | VALIDATOR, LEAD — and NOT this task's last_mapper |
        | `completed` | (none)                                            |

        Other rules:
          - No other active lock on this task.
          - Caller holds no other active lock in this project (one-lock-
            per-project rule). 409 response includes `existingLock`.
          - Project status is `open`.

        Side effects:
          - INSERT `tasking_locks` with
            `expires_at = NOW() + project.lock_timeout_hours`.
          - Emit `task_locked` audit event.
      operationId: lockTask
      responses:
        "200":
          description: Lock acquired. Body is the updated task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Role does not permit locking in this status, or self-validation guard hit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          { $ref: "#/components/responses/TaskOrProjectOrWorkspaceNotFound" }
        "409":
          description: |
            One of:
              - "Task is already locked".
              - "User already holds a lock in this project" — body includes `existingLock`.
              - "Project is not open".
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LockConflictError" }
    delete:
      tags: [locks]
      summary: Release a lock on a task.
      description: |
        Default: caller must be the active lock holder
        (`release_reason = manual`).

        With `?force=true`: caller must be LEAD
        (`release_reason = lead_release`).

        Task `status` is unchanged in both modes. Force-release does
        not relock — LEAD must POST `/lock` separately to take over.
      operationId: unlockTask
      parameters:
        - in: query
          name: force
          schema: { type: boolean, default: false }
      responses:
        "204":
          description: Released.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Not the lock holder (default mode) or not LEAD (force mode).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          { $ref: "#/components/responses/TaskOrProjectOrWorkspaceNotFound" }
        "409":
          description: Task has no active lock.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /workspaces/{workspace_id}/tasking/projects/{project_id}/tasks/{task_number}/extend:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
      - $ref: "#/components/parameters/TaskNumberPath"
    post:
      tags: [locks]
      summary: Extend the expiry of your own lock.
      description: |
        Caller must hold the active lock. Adds the project's
        `lock_timeout_hours` to the current `expires_at` (slides from
        the existing expiry, not from `NOW()`). Emits
        `task_lock_extended`.
      operationId: extendTaskLock
      responses:
        "200":
          description: Lock extended.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Caller does not hold the active lock.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          { $ref: "#/components/responses/TaskOrProjectOrWorkspaceNotFound" }
        "409":
          description: Task has no active lock to extend.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /workspaces/{workspace_id}/tasking/projects/{project_id}/tasks/{task_number}/reset:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
      - $ref: "#/components/parameters/TaskNumberPath"
    post:
      tags: [locks]
      summary: Reset a single task back to `to_map`.
      description: |
        LEAD only. Restarts work on a task — useful when a contributor
        or validator pushed a task into a wrong terminal state and
        someone needs to redo it.

        Effect (single transaction):
          - If an active lock exists, it is released with
            `release_reason = 'reset'` and a `task_unlocked` audit
            event is emitted.
          - Task `status` is set to `to_map`, `last_mapper_id` is
            cleared, and a `task_state_changed` audit event is emitted
            (only when the status actually changed).
          - A `task_reset` audit event is emitted.

        Existing changeset rows and remap-feedback rows are NOT
        deleted — only the live task state is rewound.

        Pre-conditions: project status is `open` or `done`.
      operationId: resetTask
      responses:
        "200":
          description: Task reset.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ForbiddenLeadRequired" }
        "404":
          { $ref: "#/components/responses/TaskOrProjectOrWorkspaceNotFound" }
        "422":
          description: Project is in `draft` (no tasks to reset yet).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /workspaces/{workspace_id}/tasking/projects/{project_id}/tasks/{task_number}/submit:
    parameters:
      - $ref: "#/components/parameters/WorkspaceIdPath"
      - $ref: "#/components/parameters/ProjectIdPath"
      - $ref: "#/components/parameters/TaskNumberPath"
    post:
      tags: [submit]
      summary: Submit an OSM changeset and optionally transition state.
      description: |
        The "Done?" flow. Caller must hold the active lock.

        Effective actor role is inferred from current task status:
          - `to_map` / `to_remap` → mapper context
          - `to_review`           → validator context
          (LEAD may act in either context.)

        State transition (when `done:true`):

        | Current     | Effective role     | Feedback? | New status                                              |
        |-------------|--------------------|-----------|---------------------------------------------------------|
        | `to_map`    | contributor / lead | n/a       | `to_review` if `review_required` else `completed`        |
        | `to_remap`  | contributor / lead | n/a       | `to_review`                                              |
        | `to_review` | validator / lead   | no        | `completed`                                              |
        | `to_review` | validator / lead   | yes       | `to_remap` (and INSERT `tasking_feedback` with `reasonCategory` set) |

        With `done:false` the state is unchanged but the lock's
        `expires_at` slides forward to `submitted_at + lock_timeout_hours`
        (emits `task_lock_renewed`).

        `osmChangesetId` is supplied by the client. In practice it is
        auto-populated by the editor (Rapid in the iframe) on save —
        the UI does not ask the user to type it. The id may optionally
        be verified by looking it up in the internal `osm-web` service
        (same network); a verification miss returns `422` with detail
        "Invalid OSM changeset id" and nothing is written.

        `last_mapper_id` is updated only when the effective role is
        mapper.

        Feedback rules:
          - `feedback` is meaningful only in validator context on
            `to_review`. Otherwise its presence returns 422.
          - For `to_review` → `to_remap`, `feedback.reasonCategory` is
            required (422 if missing).
      operationId: submitTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitRequest" }
      responses:
        "200":
          description: Submission accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Caller does not hold the active lock.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          { $ref: "#/components/responses/TaskOrProjectOrWorkspaceNotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }

  # =========================================================================
  # Roles
  # =========================================================================

  # NOTE: the workspace-level role surface below mirrors the shipped
  # endpoints on the `roles` branch (api/src/users/routes.py). It uses
  # `/workspaces/{workspace_id}/users/...`, not `/.../roles/...`. Only
  # privileged role rows (LEAD, VALIDATOR) are stored; CONTRIBUTOR is
  # implicit for any project-group member and cannot be assigned via PUT.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceIdPath:
      name: workspace_id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    ProjectIdPath:
      name: project_id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    TaskNumberPath:
      name: task_number
      in: path
      required: true
      description: Sequential, human-readable id scoped to the project. Starts at 1.
      schema: { type: integer, minimum: 1 }
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: Client-generated unique value (UUID recommended). Scoped per project.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    ProjectStatus:
      type: string
      enum: [draft, open, done]
    TaskStatus:
      type: string
      enum: [to_map, to_review, to_remap, completed]
...
    TaskBoundariesFeatureCollection:
      type: object
      required: [type, features]
      properties:
        type: { type: string, enum: [FeatureCollection] }
        features:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/TaskBoundaryFeature" }

    # ---------- Projects ----------
    Project:
      type: object
      required:
        - id
        - workspaceId
        - name
        - status
        - reviewRequired
        - lockTimeoutHours
        - createdBy
        - createdAt
        - updatedAt
        - hasAoi
        - taskCount
      properties:
        id: { type: integer, format: int64 }
        workspaceId: { type: integer, format: int64 }
        name: { type: string, maxLength: 255 }
        instructions: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        reviewRequired: { type: boolean, default: true }
        lockTimeoutHours:
          { type: integer, minimum: 1, maximum: 720, default: 8 }
        taskBoundaryType: { $ref: "#/components/schemas/TaskBoundaryType" }
        hasAoi: { type: boolean }
        taskCount: { type: integer, minimum: 0 }
        createdBy: { type: string, format: uuid }
        createdByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ProjectUpdateRequest:
      type: object
      description: All fields optional. Server enforces per-status mutability.
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        instructions: { type: string, maxLength: 10000, nullable: true }
        lockTimeoutHours: { type: integer, minimum: 1, maximum: 720 }
        reviewRequired: { type: boolean }

    # ---------- Tasks ----------
...
    TaskListResponse:
      type: object
      required: [tasks, pagination]
      properties:
        tasks:
          type: array
          items: { $ref: "#/components/schemas/Task" }
        pagination: { $ref: "#/components/schemas/Pagination" }
...
    ValidatePreviewResponse:
      type: object
      required: [valid, warnings, source, featureCollection]
      properties:
        valid:
          type: boolean
          description: True when no hard failures. Warnings do not affect this.
        warnings:
          type: array
          items: { $ref: "#/components/schemas/ValidateWarning" }
        source:
          $ref: "#/components/schemas/GridSource"
          description: Always `import` for this response. Pass through unchanged to `/save`.
        featureCollection:
          $ref: "#/components/schemas/TaskBoundariesFeatureCollection"
    SaveTasksRequest:
      type: object
      required: [source, featureCollection]
      properties:
        source:
          $ref: "#/components/schemas/GridSource"
          description: Sets `tasking_projects.task_boundary_type` for the project.
        featureCollection:
          $ref: "#/components/schemas/TaskBoundariesFeatureCollection"
    SaveTasksResponse:
      type: object
      required: [projectId, taskBoundaryType, taskCount, tasks]
      properties:
        projectId: { type: integer, format: int64 }
        taskBoundaryType: { $ref: "#/components/schemas/GridSource" }
        taskCount:
          type: integer
          minimum: 1
          description: Number of tasks created (== `featureCollection.features.length`).
        tasks:
          type: array
          description: All created tasks in the same order as the request's `features[]`. `taskNumber` is sequential starting at 1.
          items: { $ref: "#/components/schemas/Task" }
        idempotencyKey:
          type: string
          nullable: true
        replayed:
          type: boolean
          default: false
          description: True on idempotent replay (HTTP 200 case).

    # ---------- Submit / Lock ----------
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items: { type: object, additionalProperties: true }
          description: FastAPI default error shape — string for raised `HTTPException`, structured list for pydantic validation errors.
  responses:
    BadRequest:
      description: Malformed JSON or schema-level failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ForbiddenLeadRequired:
      description: Caller is a workspace contributor but does not hold LEAD.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ProjectOrWorkspaceNotFound:
      description: Workspace or project does not exist, or is outside the caller's tenancy.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TaskOrProjectOrWorkspaceNotFound:
      description: Workspace, project, or task does not exist, or is outside the caller's tenancy.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnprocessableEntity:
      description: Well-formed request that violates a business rule.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }